+++ src/LLVM.jl
@@ -40,9 +40,9 @@ let
     if version().major < 15
         error("LLVM.jl only supports LLVM 15 and later.")
     end
-    dir = if version().major > 21
-        @warn "LLVM.jl has not been tested with LLVM versions newer than 21."
-        joinpath(@__DIR__, "..", "lib", "21")
+    dir = if version().major > 22
+        @warn "LLVM.jl has not been tested with LLVM versions newer than 22."
+        joinpath(@__DIR__, "..", "lib", "22")
     else
         joinpath(@__DIR__, "..", "lib", string(version().major))
     end